+++ src/components/ShippingAddressForm/ShippingAddressForm.tsx
@@ -1,0 +1,126 @@
+import "./scss/index.scss";
+
+import * as React from "react";
+
+import { Button, Form, TextField, SelectField } from "..";
+import { ShopContext } from "../ShopProvider/context";
+import { IShippingAddressFormProps } from "./types";
+import { getFormData } from "./utils";
+
+const ShippingAddressForm: React.SFC<IShippingAddressFormProps> = ({
+  data,
+  billing,
+  buttonText,
+  errors,
+  loading,
+  onSubmit,
+  children
+}) => (
+  <div className="address-form">
+    <ShopContext.Consumer>
+      {({ countries, geolocalization, defaultCountry }) => (
+        <Form
+          errors={errors}
+          onSubmit={(evt, data) => onSubmit(evt, data as any)}
+          data={getFormData(geolocalization, defaultCountry, data)}
+        >
+          {children}
+          <div className="address-form__grid">
+            <TextField
+              label="First Name"
+              type="given-name"
+              name="firstName"
+              autoComplete="given-name"
+              required
+            />
+            <TextField
+              label="Last Name"
+              type="family-name"
+              name="lastName"
+              autoComplete="family-name"
+              required
+            />
+          </div>
+          <TextField
+            label="Street Line 1"
+            type="address-line1"
+            name="streetAddress1"
+            autoComplete="address-line1"
+            required
+          />
+          <TextField
+            label="Street Line 2"
+            type="address-line2"
+            name="streetAddress2"
+            autoComplete="address-line2"
+          />
+          <TextField
+            label="Company"
+            type="organization"
+            name="companyName"
+            autoComplete="organization"
+          />
+          <div className="address-form__grid">
+            <TextField
+              label="City"
+              type="city"
+              name="city"
+              autoComplete="address-level2"
+            />
+            <TextField
+              label="State/Province"
+              type="state"
+              name="countryArea"
+              autoComplete="address-level1"
+              required
+            />
+          </div>
+          <div className="address-form__grid">
+            <TextField
+              label="Zip-Code"
+              type="postal-code"
+              name="postalCode"
+              autoComplete="postal-code"
+              required
+            />
+            <SelectField
+              label="Country"
+              name="country"
+              options={countries.map(country => ({
+                label: country.country,
+                value: country.code
+              }))}
+            />
+          </div>
+          {!billing && (
+            <TextField
+              label="Email Address"
+              type="email"
+              autoComplete="email"
+              name="email"
+              required
+            />
+          )}
+          <TextField
+            label="Phone number"
+            type="tel"
+            name="phone"
+            autoComplete="tel"
+          />
+          {!billing ? (
+            <label className="checkbox">
+              <input name="asBilling" type="checkbox" />
+              <span>Use as Billing Address</span>
+            </label>
+          ) : null}
+
+          <Button type="submit" disabled={loading}>
+            {loading ? "Loading" : buttonText}
+          </Button>
+        </Form>
+      )}
+    </ShopContext.Consumer>
+  </div>
+);
+
+export default ShippingAddressForm;
